fix(storage): installed couchdb-find plugin to support Mongo-like `find()` operator
when CouchDB saves a document, it assign it an `_id` field which it generates, which is different
than the `id` field that I'm generating when I'm creating a task. When `updateTask` is called, it
gets the task's id but CouchDB needs its `_id` field which is not present on the task, so I need to
make a Mongo-like query (`find`) to find my task and retrieve its `_id` field and couchdb-find is a
plugin which support this kind of queries.